Add refresh request details after action modal in the list
