+++ src/smart-components/request/requests.js
@@ -1,7 +1,4 @@
-import React, { Fragment, useContext } from 'react';
-import { Route } from 'react-router-dom';
-import { fetchRequests } from '../../redux/actions/request-actions';
-import ActionModal from './action-modal';
+import React, { useContext } from 'react';
 import {
   APPROVAL_APPROVER_PERSONA, useIsApprovalAdmin,
   useIsApprovalApprover,
@@ -18,16 +15,6 @@ const Requests = () => {
   const { userRoles: userRoles } = useContext(UserContext);
   const isApprovalAdmin = useIsApprovalAdmin(userRoles);
   const isApprovalApprover = useIsApprovalApprover(userRoles);
-
-  const routes = () => <Fragment>
-    <Route exact path={ routesLinks.requests.addComment } render={ props => <ActionModal { ...props }
-      actionType={ 'Add Comment' }
-      postMethod={ fetchRequests } /> }/>
-    <Route exact path={ routesLinks.requests.approve } render={ props => <ActionModal { ...props } actionType={ 'Approve' }
-      postMethod={ fetchRequests }/> } />
-    <Route exact path={ routesLinks.requests.deny } render={ props => <ActionModal { ...props } actionType={ 'Deny' }
-      postMethod={ fetchRequests }/> } />
-  </Fragment>;
 
   const actionsDisabled = (requestData) => requestData &&
     requestData.state ?
@@ -48,7 +35,6 @@ const Requests = () => {
   return !isApprovalApprover ?
     <EmptyRequestList/>
     : <RequestsList
-      routes={ routes }
       persona={ APPROVAL_APPROVER_PERSONA }
       actionResolver={ actionResolver }
     />;
